:zap: Refactoring split & draggable -> modules
