--- packages/draggable/babel.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  presets: [
-    '@vue/app'
-  ]
-}
