--- packages/draggable/index.ts
@@ -1,1 +1,0 @@
-export { DragElement as default, Draggable } from './src';
