--- packages/split/package.json
@@ -1,76 +1,0 @@
-{
-  "name": "@dawgjs/split",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "storybook:build": "vue-cli-service storybook:build -c config/storybook",
-    "storybook:serve": "vue-cli-service storybook:serve -p 6006 -c config/storybook",
-    "test:unit": "vue-cli-service test:unit"
-  },
-  "dependencies": {
-    "vue-class-component": "^6.0.0",
-    "vue-property-decorator": "^7.0.0",
-    "@dawgjs/draggable": "^0.1.0"
-  },
-  "devDependencies": {
-    "@storybook/addon-actions": "^4.1.0",
-    "@storybook/addon-knobs": "^4.1.0",
-    "@storybook/addon-links": "^4.1.0",
-    "@storybook/addon-notes": "^4.1.0",
-    "@types/jest": "^23.1.4",
-    "@vue/cli-plugin-babel": "^3.4.0",
-    "@vue/cli-plugin-typescript": "^3.4.0",
-    "@vue/cli-plugin-unit-jest": "^3.4.0",
-    "@vue/cli-service": "^3.4.0",
-    "@vue/test-utils": "^1.0.0-beta.20",
-    "babel-core": "7.0.0-bridge.0",
-    "ts-jest": "^23.0.0",
-    "typescript": "^3.0.0",
-    "vue": "^2.5.17",
-    "vue-cli-plugin-storybook": "^0.5.2",
-    "vue-template-compiler": "^2.5.21"
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ],
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "jsx",
-      "json",
-      "vue",
-      "ts",
-      "tsx"
-    ],
-    "transform": {
-      "^.+\\.vue$": "vue-jest",
-      ".+\\.(css|styl|less|sass|scss|svg|png|jpg|ttf|woff|woff2)$": "jest-transform-stub",
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "moduleNameMapper": {
-      "^@/(.*)$": "<rootDir>/src/$1"
-    },
-    "snapshotSerializers": [
-      "jest-serializer-vue"
-    ],
-    "testMatch": [
-      "**/src/**/*.spec.(js|ts)"
-    ],
-    "testURL": "http://localhost/",
-    "globals": {
-      "ts-jest": {
-        "babelConfig": true
-      }
-    }
-  }
-}
